requirements.txt
opencv-python
numpy
pillow
tqdm